move blender plugins to host folder
