+++ pype/hosts/blender/plugins/create/create_animation.py
@@ -3,7 +3,7 @@
 import bpy
 
 from avalon import api, blender
-import pype.hosts.blender.plugin
+import pype.hosts.blender.api.plugin
 
 
 class CreateAnimation(blender.Creator):
@@ -17,7 +17,7 @@ class CreateAnimation(blender.Creator):
     def process(self):
         asset = self.data["asset"]
         subset = self.data["subset"]
-        name = pype.hosts.blender.plugin.asset_name(asset, subset)
+        name = pype.hosts.blender.api.plugin.asset_name(asset, subset)
         collection = bpy.data.collections.new(name=name)
         bpy.context.scene.collection.children.link(collection)
         self.data['task'] = api.Session.get('AVALON_TASK')
