+++ pype/hosts/blender/plugins/create/create_camera.py
@@ -1,25 +1,25 @@
-"""Create a model asset."""
+"""Create a camera asset."""
 
 import bpy
 
 from avalon import api
 from avalon.blender import Creator, lib
-import pype.hosts.blender.plugin
+import pype.hosts.blender.api.plugin
 
 
-class CreateModel(Creator):
+class CreateCamera(Creator):
     """Polygonal static geometry"""
 
-    name = "modelMain"
-    label = "Model"
-    family = "model"
-    icon = "cube"
+    name = "cameraMain"
+    label = "Camera"
+    family = "camera"
+    icon = "video-camera"
 
     def process(self):
 
         asset = self.data["asset"]
         subset = self.data["subset"]
-        name = pype.hosts.blender.plugin.asset_name(asset, subset)
+        name = pype.hosts.blender.api.plugin.asset_name(asset, subset)
         collection = bpy.data.collections.new(name=name)
         bpy.context.scene.collection.children.link(collection)
         self.data['task'] = api.Session.get('AVALON_TASK')
