+++ pype/hosts/blender/plugins/create/create_layout.py
@@ -4,7 +4,7 @@ import bpy
 
 from avalon import api
 from avalon.blender import Creator, lib
-import pype.hosts.blender.plugin
+import pype.hosts.blender.api.plugin
 
 
 class CreateLayout(Creator):
@@ -19,7 +19,7 @@ class CreateLayout(Creator):
 
         asset = self.data["asset"]
         subset = self.data["subset"]
-        name = pype.hosts.blender.plugin.asset_name(asset, subset)
+        name = pype.hosts.blender.api.plugin.asset_name(asset, subset)
         collection = bpy.data.collections.new(name=name)
         bpy.context.scene.collection.children.link(collection)
         self.data['task'] = api.Session.get('AVALON_TASK')
