+++ pype/hosts/blender/plugins/load/load_animation.py
@@ -7,14 +7,14 @@ from typing import Dict, List, Optional
 
 from avalon import api, blender
 import bpy
-import pype.hosts.blender.plugin
+import pype.hosts.blender.api.plugin
 
 
 logger = logging.getLogger("pype").getChild(
     "blender").getChild("load_animation")
 
 
-class BlendAnimationLoader(pype.hosts.blender.plugin.AssetLoader):
+class BlendAnimationLoader(pype.hosts.blender.api.plugin.AssetLoader):
     """Load animations from a .blend file.
 
     Warning:
@@ -105,8 +105,8 @@ class BlendAnimationLoader(pype.hosts.blender.plugin.AssetLoader):
         libpath = self.fname
         asset = context["asset"]["name"]
         subset = context["subset"]["name"]
-        lib_container = pype.hosts.blender.plugin.asset_name(asset, subset)
-        container_name = pype.hosts.blender.plugin.asset_name(
+        lib_container = pype.hosts.blender.api.plugin.asset_name(asset, subset)
+        container_name = pype.hosts.blender.api.plugin.asset_name(
             asset, subset, namespace
         )
 
@@ -175,7 +175,7 @@ class BlendAnimationLoader(pype.hosts.blender.plugin.AssetLoader):
         assert libpath.is_file(), (
             f"The file doesn't exist: {libpath}"
         )
-        assert extension in pype.hosts.blender.plugin.VALID_EXTENSIONS, (
+        assert extension in pype.hosts.blender.api.plugin.VALID_EXTENSIONS, (
             f"Unsupported file: {libpath}"
         )
