+++ pype/hosts/blender/plugins/load/load_model.py
@@ -1,4 +1,4 @@
-"""Load a camera asset in Blender."""
+"""Load a model asset in Blender."""
 
 import logging
 from pathlib import Path
@@ -7,75 +7,69 @@ from typing import Dict, List, Optional
 
 from avalon import api, blender
 import bpy
-import pype.hosts.blender.plugin
-
-logger = logging.getLogger("pype").getChild("blender").getChild("load_camera")
+import pype.hosts.blender.api.plugin as plugin
 
 
-class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
-    """Load a camera from a .blend file.
+class BlendModelLoader(plugin.AssetLoader):
+    """Load models from a .blend file.
 
-    Warning:
-        Loading the same asset more then once is not properly supported at the
-        moment.
+    Because they come from a .blend file we can simply link the collection that
+    contains the model. There is no further need to 'containerise' it.
     """
 
-    families = ["camera"]
+    families = ["model"]
     representations = ["blend"]
 
-    label = "Link Camera"
+    label = "Link Model"
     icon = "code-fork"
     color = "orange"
 
-    def _remove(self, objects, lib_container):
+    def _remove(self, objects, container):
         for obj in list(objects):
-            bpy.data.cameras.remove(obj.data)
+            for material_slot in list(obj.material_slots):
+                bpy.data.materials.remove(material_slot.material)
+            bpy.data.meshes.remove(obj.data)
 
-        bpy.data.collections.remove(bpy.data.collections[lib_container])
+        bpy.data.collections.remove(container)
 
-    def _process(self, libpath, lib_container, container_name, actions):
-
+    def _process(
+        self, libpath, lib_container, container_name,
+        parent_collection
+    ):
         relative = bpy.context.preferences.filepaths.use_relative_paths
         with bpy.data.libraries.load(
             libpath, link=True, relative=relative
         ) as (_, data_to):
             data_to.collections = [lib_container]
 
-        scene = bpy.context.scene
+        parent = parent_collection
 
-        scene.collection.children.link(bpy.data.collections[lib_container])
+        if parent is None:
+            parent = bpy.context.scene.collection
 
-        camera_container = scene.collection.children[lib_container].make_local()
+        parent.children.link(bpy.data.collections[lib_container])
 
-        objects_list = []
+        model_container = parent.children[lib_container].make_local()
+        model_container.name = container_name
 
-        for obj in camera_container.objects:
-            local_obj = obj.make_local()
-            local_obj.data.make_local()
+        for obj in model_container.objects:
+            local_obj = plugin.prepare_data(obj, container_name)
+            plugin.prepare_data(local_obj.data, container_name)
 
-            if not local_obj.get(blender.pipeline.AVALON_PROPERTY):
+            for material_slot in local_obj.material_slots:
+                plugin.prepare_data(material_slot.material, container_name)
+
+            if not obj.get(blender.pipeline.AVALON_PROPERTY):
                 local_obj[blender.pipeline.AVALON_PROPERTY] = dict()
 
             avalon_info = local_obj[blender.pipeline.AVALON_PROPERTY]
             avalon_info.update({"container_name": container_name})
 
-            if actions[0] is not None:
-                if local_obj.animation_data is None:
-                    local_obj.animation_data_create()
-                local_obj.animation_data.action = actions[0]
-
-            if actions[1] is not None:
-                if local_obj.data.animation_data is None:
-                    local_obj.data.animation_data_create()
-                local_obj.data.animation_data.action = actions[1]
-
-            objects_list.append(local_obj)
-
-        camera_container.pop(blender.pipeline.AVALON_PROPERTY)
+        model_container.pop(blender.pipeline.AVALON_PROPERTY)
 
         bpy.ops.object.select_all(action='DESELECT')
 
-        return objects_list
+        return model_container
 
     def process_asset(
         self, context: dict, name: str, namespace: Optional[str] = None,
@@ -92,9 +86,16 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         libpath = self.fname
         asset = context["asset"]["name"]
         subset = context["subset"]["name"]
-        lib_container = pype.hosts.blender.plugin.asset_name(asset, subset)
-        container_name = pype.hosts.blender.plugin.asset_name(
-            asset, subset, namespace
+
+        lib_container = plugin.asset_name(
+            asset, subset
+        )
+        unique_number = plugin.get_unique_number(
+            asset, subset
+        )
+        namespace = namespace or f"{asset}_{unique_number}"
+        container_name = plugin.asset_name(
+            asset, subset, unique_number
         )
 
         container = bpy.data.collections.new(lib_container)
@@ -113,11 +114,13 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         container_metadata["libpath"] = libpath
         container_metadata["lib_container"] = lib_container
 
-        objects_list = self._process(
-            libpath, lib_container, container_name, (None, None))
+        obj_container = self._process(
+            libpath, lib_container, container_name, None)
+
+        container_metadata["obj_container"] = obj_container
 
         # Save the list of objects in the metadata container
-        container_metadata["objects"] = objects_list
+        container_metadata["objects"] = obj_container.all_objects
 
         nodes = list(container.objects)
         nodes.append(container)
@@ -136,15 +139,13 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         Warning:
             No nested collections are supported at the moment!
         """
-
         collection = bpy.data.collections.get(
             container["objectName"]
         )
-
         libpath = Path(api.get_representation_path(representation))
         extension = libpath.suffix.lower()
 
-        logger.info(
+        self.log.info(
             "Container: %s\nRepresentation: %s",
             pformat(container, indent=2),
             pformat(representation, indent=2),
@@ -162,15 +163,21 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         assert libpath.is_file(), (
             f"The file doesn't exist: {libpath}"
         )
-        assert extension in pype.hosts.blender.plugin.VALID_EXTENSIONS, (
+        assert extension in plugin.VALID_EXTENSIONS, (
             f"Unsupported file: {libpath}"
         )
 
         collection_metadata = collection.get(
             blender.pipeline.AVALON_PROPERTY)
         collection_libpath = collection_metadata["libpath"]
-        objects = collection_metadata["objects"]
         lib_container = collection_metadata["lib_container"]
+
+        obj_container = plugin.get_local_collection_with_name(
+            collection_metadata["obj_container"].name
+        )
+        objects = obj_container.all_objects
+
+        container_name = obj_container.name
 
         normalized_collection_libpath = (
             str(Path(bpy.path.abspath(collection_libpath)).resolve())
@@ -178,39 +185,27 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         normalized_libpath = (
             str(Path(bpy.path.abspath(str(libpath))).resolve())
         )
-        logger.debug(
+        self.log.debug(
             "normalized_collection_libpath:\n  %s\nnormalized_libpath:\n  %s",
             normalized_collection_libpath,
             normalized_libpath,
         )
         if normalized_collection_libpath == normalized_libpath:
-            logger.info("Library already loaded, not updating...")
+            self.log.info("Library already loaded, not updating...")
             return
 
-        camera = objects[0]
+        parent = plugin.get_parent_collection(obj_container)
 
-        camera_action = None
-        camera_data_action = None
+        self._remove(objects, obj_container)
 
-        if camera.animation_data and camera.animation_data.action:
-            camera_action = camera.animation_data.action
-
-        if camera.data.animation_data and camera.data.animation_data.action:
-            camera_data_action = camera.data.animation_data.action
-
-        actions = (camera_action, camera_data_action)
-
-        self._remove(objects, lib_container)
-
-        objects_list = self._process(
-            str(libpath), lib_container, collection.name, actions)
+        obj_container = self._process(
+            str(libpath), lib_container, container_name, parent)
 
         # Save the list of objects in the metadata container
-        collection_metadata["objects"] = objects_list
+        collection_metadata["obj_container"] = obj_container
+        collection_metadata["objects"] = obj_container.all_objects
         collection_metadata["libpath"] = str(libpath)
         collection_metadata["representation"] = str(representation["_id"])
-
-        bpy.ops.object.select_all(action='DESELECT')
 
     def remove(self, container: Dict) -> bool:
         """Remove an existing container from a Blender scene.
@@ -225,7 +220,6 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
         Warning:
             No nested collections are supported at the moment!
         """
-
         collection = bpy.data.collections.get(
             container["objectName"]
         )
@@ -237,11 +231,76 @@ class BlendCameraLoader(pype.hosts.blender.plugin.AssetLoader):
 
         collection_metadata = collection.get(
             blender.pipeline.AVALON_PROPERTY)
-        objects = collection_metadata["objects"]
-        lib_container = collection_metadata["lib_container"]
 
-        self._remove(objects, lib_container)
+        obj_container = plugin.get_local_collection_with_name(
+            collection_metadata["obj_container"].name
+        )
+        objects = obj_container.all_objects
+
+        self._remove(objects, obj_container)
 
         bpy.data.collections.remove(collection)
 
         return True
+
+
+class CacheModelLoader(plugin.AssetLoader):
+    """Load cache models.
+
+    Stores the imported asset in a collection named after the asset.
+
+    Note:
+        At least for now it only supports Alembic files.
+    """
+
+    families = ["model"]
+    representations = ["abc"]
+
+    label = "Link Model"
+    icon = "code-fork"
+    color = "orange"
+
+    def process_asset(
+        self, context: dict, name: str, namespace: Optional[str] = None,
+        options: Optional[Dict] = None
+    ) -> Optional[List]:
+        """
+        Arguments:
+            name: Use pre-defined name
+            namespace: Use pre-defined namespace
+            context: Full parenthood of representation to load
+            options: Additional settings dictionary
+        """
+        raise NotImplementedError(
+            "Loading of Alembic files is not yet implemented.")
+        # TODO (jasper): implement Alembic import.
+
+        libpath = self.fname
+        asset = context["asset"]["name"]
+        subset = context["subset"]["name"]
+        # TODO (jasper): evaluate use of namespace which is 'alien' to Blender.
+        lib_container = container_name = (
+            plugin.asset_name(asset, subset, namespace)
+        )
+        relative = bpy.context.preferences.filepaths.use_relative_paths
+
+        with bpy.data.libraries.load(
+            libpath, link=True, relative=relative
+        ) as (data_from, data_to):
+            data_to.collections = [lib_container]
+
+        scene = bpy.context.scene
+        instance_empty = bpy.data.objects.new(
+            container_name, None
+        )
+        scene.collection.objects.link(instance_empty)
+        instance_empty.instance_type = 'COLLECTION'
+        collection = bpy.data.collections[lib_container]
+        collection.name = container_name
+        instance_empty.instance_collection = collection
+
+        nodes = list(collection.objects)
+        nodes.append(collection)
+        nodes.append(instance_empty)
+        self[:] = nodes
+        return nodes
