+++ pype/hosts/blender/plugins/load/load_rig.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional
 
 from avalon import api, blender
 import bpy
-import pype.hosts.blender.plugin as plugin
+import pype.hosts.blender.api.plugin as plugin
 
 
 class BlendRigLoader(plugin.AssetLoader):
