+++ pype/hosts/blender/plugins/publish/collect_current_file.py
