+++ pype/hosts/blender/plugins/publish/collect_instances.py
