+++ pype/hosts/blender/plugins/publish/extract_abc.py
@@ -1,7 +1,7 @@
 import os
 
 import pype.api
-import pype.hosts.blender.plugin
+import pype.hosts.blender.api.plugin
 
 import bpy
 
@@ -61,7 +61,8 @@ class ExtractABC(pype.api.Extractor):
             except:
                 continue
 
-        new_context = pype.hosts.blender.plugin.create_blender_context(active=selected[0], selected=selected)
+        new_context = pype.hosts.blender.api.plugin.create_blender_context(
+            active=selected[0], selected=selected)
 
         # We set the scale of the scene for the export
         scene.unit_settings.scale_length = 0.01
