+++ pype/hosts/blender/plugins/publish/extract_animation_collection.py
@@ -6,6 +6,7 @@ import pyblish.api
 
 import bpy
 
+
 class ExtractSetDress(pype.api.Extractor):
     """Extract setdress."""
 
@@ -21,17 +22,20 @@ class ExtractSetDress(pype.api.Extractor):
         json_data = []
 
         for i in instance.context:
-            collection = i.data.get('name')
+            collection = i.data.get("name")
             container = None
             for obj in bpy.data.collections[collection].objects:
-                if obj.type == 'ARMATURE':
-                    container_name = obj.get('avalon').get('container_name')
+                if obj.type == "ARMATURE":
+                    container_name = obj.get("avalon").get("container_name")
                     container = bpy.data.collections[container_name]
             if container:
-                json_dict = {}
-                json_dict['subset'] = i.data.get('subset')
-                json_dict['container'] = container.name
-                json_dict['instance_name'] = container.get('avalon').get('instance_name')
+                json_dict = {
+                    "subset": i.data.get("subset"),
+                    "container": container.name,
+                }
+                json_dict["instance_name"] = container.get("avalon").get(
+                    "instance_name"
+                )
                 json_data.append(json_dict)
 
         if "representations" not in instance.data:
@@ -44,13 +48,14 @@ class ExtractSetDress(pype.api.Extractor):
             json.dump(json_data, fp=file, indent=2)
 
         json_representation = {
-            'name': 'json',
-            'ext': 'json',
-            'files': json_filename,
+            "name": "json",
+            "ext": "json",
+            "files": json_filename,
             "stagingDir": stagingdir,
         }
         instance.data["representations"].append(json_representation)
 
-        self.log.info("Extracted instance '{}' to: {}".format(
-                      instance.name, json_representation))
-
+        self.log.info(
+            "Extracted instance '{}' to: {}".format(instance.name,
+                                                    json_representation)
+        )
