+++ pype/hosts/blender/plugins/publish/extract_fbx.py
