+++ pype/hosts/blender/plugins/publish/extract_fbx_animation.py
