+++ pype/hosts/blender/plugins/publish/integrate_animation.py
